added composer support
