+++ src/RussianPostTrackingException.php
@@ -1,0 +1,3 @@
+<?php
+
+class RussianPostTrackingException extends Exception {}
